TouchpadInputMapper: dedup successive swipe lift gestures in metrics
In tests, the Gestures library occasionally outputs two lift gestures in
a row, which can cause inaccurate metrics reporting. To work around
this, keep track of the previous gesture the metrics collector saw, and
don't report additional swipe gestures if we see multiple lifts in a
row.

Bug: 397219482
Test: $ atest \
        CtsInputHostTestCases:android.input.cts.hostside.InputAtomsTest
Flag: TEST_ONLY
Change-Id: Iac2b6ae68dc62e3a106a743b667eddd96a1a7a74